requirements.txt
django==4.2.13
psycopg2-binary==2.9.9
django-encrypted-model-fields
django-axes==6.4.0
django-otp==1.3.0
qrcode==7.4.2
celery==5.3.6
redis==5.0.1
reportlab==4.1.0
openpyxl==3.1.2
django-crispy-forms==2.1
crispy-bootstrap5==0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==10.2.0
django-ratelimit==4.1.0
pytest-django==4.7.0
factory-boy==3.3.0
django-debug-toolbar==4.3.0
